Terminal_Escape_Web/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
gunicorn==23.0.0   